fixed buy/sell bug
Previously, stockpage.handlebars would reference the dropdown field in order to get the user id. when the dropdown was removed, this was not updated.

now, the stockpages.handlebars gets the userid from sessionstorage.

Also, the script to add the options to the userid dropdown was not removed from main.handlebars.

as a result, the script was running twice since the script was ported over to login.handlebars. i removed this redundant script.